PIDMR-232 Restructure components into dedicated directories
