+++ src/main/components/main.tsx
@@ -1,8 +1,8 @@
 import ReactDOM from "react-dom/client";
-import App from "./App.tsx";
-import "bootstrap/dist/css/bootstrap.min.css";
-import "./index.css";
 import { BrowserRouter } from "react-router-dom";
+import "bootstrap/dist/css/bootstrap.min.css";
+import "../styles/index.css";
+import App from "./App";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <BrowserRouter>
