+++ src/pages/about/Privacy.tsx
