+++ src/pages/about/Terms.tsx
