+++ src/pages/supported-pids/SupportedPids.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { useSearchParams, useNavigate, Link } from "react-router-dom";
-import { ApiResponse } from "./types";
+import { ApiResponse } from "../../types";
 
 // API endpoint declared in env variable
 const PIDMR_API = import.meta.env.VITE_PIDMR_API;
